Wrap formatting methods in module
