+++ lib/geordi.rb
@@ -1,2 +1,1 @@
-require 'geordi/ui'
 require 'geordi/cli'
